refactor(web): check login strictly on every page
